jobsStatusSilver Refactor
